+++ R/recomb_map_inds.R
@@ -3,7 +3,7 @@
 #' Given a list of individuals with recombination breaks given in genetic distance (such as the output of [recomb_fam()]), and a genetic map (see [`recomb_map_hg`]), this function determines all positions in base pair coordinates.
 #' If base pair positions existed in input, they are overwritten.
 #'
-#' Genetic positions are converted to base pair positions from the provided map using linear interpolation, using [stats::approx()] with options `rule = 2` (out of range cases are set to nearest end's value) and `ties = list( 'ordered', mean )` (assume data is ordered, interpolate ties in genetic distance in map using mean of base pair positions).
+#' Genetic positions are converted to base pair positions from the provided map using linear interpolation, using [stats::approx()] with options `rule = 2` (out of range cases are set to nearest end's value) and `ties = list( 'ordered', mean )` (assume data is ordered, interpolate ties in genetic position in map using mean of base pair positions).
 #' Output will be incorrect, without throwing errors, if genetic map is not ordered.
 #' Base pair positions are rounded to integers.
 #'